+++ stroom-ui/src/components/password/ResetPassword/index.ts
@@ -1,1 +1,3 @@
+import useResetPassword from "./useResetPassword";
 export { default as ResetPassword } from "./ResetPasswordContainer";
+export { useResetPassword };