+++ stroom-ui/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [v7.0-beta.5] - 2020-03-10
+
+* Fixes to for auth-into-stroom compatibility
+
+## [v7.0-beta.4] - 2020-01-14
+
+* Merged Stroom 6.0 compatibility
 
 ## [v7.0-beta.3] - 2019-12-04
 
@@ -98,7 +105,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Experimental first tag
 
-[unreleased]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.3...master
+[unreleased]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.5...master
+[v7.0-beta.5]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.4...v7.0-beta.5
+[v7.0-beta.4]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.3...v7.0-beta.4
 [v7.0-beta.3]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.2...v7.0-beta.3
 [v7.0-beta.2]: https://github.com/gchq/stroom-auth/compare/v7.0-beta.1...v7.0-beta.2
 [v7.0-beta.1]: https://github.com/gchq/stroom-auth/compare/v7.0-alpha.11...v7.0-beta.1